Redraw Lines command functional
Redraw Lines command clears the canvas of current lines and then draws again from the list of relationships.

Added _canvasPointer global to track the Point at which the next ClassBox should be drawn.

Added _classBoxCount to track the number of classes currently drawn to the screen (used to update _canvasPointer)